requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
pydantic==2.9.2
pydantic-settings==2.1.0
aiofiles==23.2.1
openpyxl==3.1.5
psycopg[binary]==3.2.9
email-validator==2.1.1
gspread==6.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
APScheduler==3.10.4
